refactor: move DB read functions to ChainStore trait
