+++ crates/store/src/state/mod.rs
@@ -1,3 +1,4 @@
+pub mod block_state_record;
 pub mod mem_state_db;
 pub mod state_db;
 mod state_tracker;
