--- crates/store/src/transaction/mem_pool_store.rs
@@ -1,94 +1,0 @@
-use gw_common::H256;
-use gw_db::{
-    error::Error,
-    schema::{
-        COLUMN_MEM_POOL_TRANSACTION, COLUMN_MEM_POOL_TRANSACTION_RECEIPT,
-        COLUMN_MEM_POOL_WITHDRAWAL,
-    },
-};
-use gw_types::{packed, prelude::*};
-
-use super::StoreTransaction;
-use crate::traits::KVStore;
-
-impl StoreTransaction {
-    pub fn insert_mem_pool_transaction(
-        &self,
-        tx_hash: &H256,
-        tx: packed::L2Transaction,
-    ) -> Result<(), Error> {
-        self.insert_raw(
-            COLUMN_MEM_POOL_TRANSACTION,
-            tx_hash.as_slice(),
-            tx.as_slice(),
-        )
-    }
-
-    pub fn get_mem_pool_transaction(
-        &self,
-        tx_hash: &H256,
-    ) -> Result<Option<packed::L2Transaction>, Error> {
-        Ok(self
-            .get(COLUMN_MEM_POOL_TRANSACTION, tx_hash.as_slice())
-            .map(|slice| {
-                packed::L2TransactionReader::from_slice_should_be_ok(slice.as_ref()).to_entity()
-            }))
-    }
-
-    pub fn remove_mem_pool_transaction(&self, tx_hash: &H256) -> Result<(), Error> {
-        self.delete(COLUMN_MEM_POOL_TRANSACTION, tx_hash.as_slice())?;
-        self.delete(COLUMN_MEM_POOL_TRANSACTION_RECEIPT, tx_hash.as_slice())?;
-        Ok(())
-    }
-
-    pub fn insert_mem_pool_transaction_receipt(
-        &self,
-        tx_hash: &H256,
-        tx_receipt: packed::TxReceipt,
-    ) -> Result<(), Error> {
-        self.insert_raw(
-            COLUMN_MEM_POOL_TRANSACTION_RECEIPT,
-            tx_hash.as_slice(),
-            tx_receipt.as_slice(),
-        )
-    }
-
-    pub fn get_mem_pool_transaction_receipt(
-        &self,
-        tx_hash: &H256,
-    ) -> Result<Option<packed::TxReceipt>, Error> {
-        Ok(self
-            .get(COLUMN_MEM_POOL_TRANSACTION_RECEIPT, tx_hash.as_slice())
-            .map(|slice| {
-                packed::TxReceiptReader::from_slice_should_be_ok(slice.as_ref()).to_entity()
-            }))
-    }
-
-    pub fn insert_mem_pool_withdrawal(
-        &self,
-        withdrawal_hash: &H256,
-        withdrawal: packed::WithdrawalRequest,
-    ) -> Result<(), Error> {
-        self.insert_raw(
-            COLUMN_MEM_POOL_WITHDRAWAL,
-            withdrawal_hash.as_slice(),
-            withdrawal.as_slice(),
-        )
-    }
-
-    pub fn get_mem_pool_withdrawal(
-        &self,
-        withdrawal_hash: &H256,
-    ) -> Result<Option<packed::WithdrawalRequest>, Error> {
-        Ok(self
-            .get(COLUMN_MEM_POOL_WITHDRAWAL, withdrawal_hash.as_slice())
-            .map(|slice| {
-                packed::WithdrawalRequestReader::from_slice_should_be_ok(slice.as_ref()).to_entity()
-            }))
-    }
-
-    pub fn remove_mem_pool_withdrawal(&self, withdrawal_hash: &H256) -> Result<(), Error> {
-        self.delete(COLUMN_MEM_POOL_WITHDRAWAL, withdrawal_hash.as_slice())?;
-        Ok(())
-    }
-}
