+++ crates/store/src/transaction/mod.rs
@@ -1,5 +1,3 @@
-pub mod mem_pool_store;
-pub mod state;
 mod store_transaction;
 
 pub use store_transaction::StoreTransaction;
